[984] Fix report issues and clean up validation files structure
