+++ terraform/pipeline/step-functions/TransformCQCData-StepFunction.json
@@ -41,7 +41,7 @@
             "Set Parameters": {
               "Type": "Pass",
               "Result": {
-                "validate_locations_script": "validate_locations_cleaned.py",
+                "validate_locations_cleaned_script": "validate_locations_cleaned.py",
                 "bucket_name_arg": "--bucket_name",
                 "bucket_name": "${dataset_bucket_name}",
                 "source_path_arg": "--source_path",
@@ -85,7 +85,7 @@
                   "ContainerOverrides": [
                     {
                       "Name": "cqc-api-container",
-                      "Command.$": "States.Array($.validate_locations_script, $.bucket_name_arg, $.bucket_name, $.source_path_arg, $.source_path, $.reports_path_arg, $.reports_path, $.compare_path_arg, $.compare_path)"
+                      "Command.$": "States.Array($.validate_locations_cleaned_script, $.bucket_name_arg, $.bucket_name, $.source_path_arg, $.source_path, $.reports_path_arg, $.reports_path, $.compare_path_arg, $.compare_path)"
                     }
                   ]
                 }
